Upgrade libraries and add a github action build
