+++ .github/workflows/deploy-dev.yaml
@@ -1,0 +1,36 @@
+name: build_and_deploy
+on: 
+  push:
+    branches:
+    - master
+    tags:
+    - '[0-9]+-[0-9]+-[0-9]+.[0-9]+'
+    paths:
+    - '**.ts'
+    - 'package.json'
+jobs:
+  deploy-dev:
+    runs-on: ubuntu-latest
+    if: github.ref == 'refs/heads/master'
+    steps: 
+    - uses: actions/checkout@v1
+    - uses: GoogleCloudPlatform/github-actions/setup-gcloud@master
+    with:
+        version: '274.0.0'
+        service_account_email: ${{ secrets.GCP_SA_EMAIL_DEV }}
+        service_account_key: ${{ secrets.GCP_SA_KEY_DEV }}
+    - run: gcloud functions deploy ${{ secrets.WEBHOOK_NAME }} --runtime nodejs10 --trigger-http
+  deploy-prod:
+    runs-on: ubuntu-latest
+    if: startsWith(github.ref, 'refs/tags/')
+    steps: 
+    - uses: actions/checkout@v1
+    - uses: GoogleCloudPlatform/github-actions/setup-gcloud@master
+    with:
+        version: '274.0.0'
+        service_account_email: ${{ secrets.GCP_SA_EMAIL_PROD }}
+        service_account_key: ${{ secrets.GCP_SA_KEY_PROD }}
+    - run: gcloud functions deploy ${{ secrets.WEBHOOK_NAME }} --runtime nodejs10 --trigger-http
+    
+    
+
