+++ package.json
@@ -1,17 +1,21 @@
 {
   "name": "@mabl/stack-driver-integration-demo",
   "description": "Mabl Demo for integrating test output with stack driver monitoring",
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/mablhq/stack-driver-integration-demo"
   },
   "license": "MIT",
   "engines": {
-    "node": "6.11.1"
+    "node": "10.12.0"
+  },
+  "scripts": {
+    "gcp-build": "tsc"
   },
   "dependencies": {
-    "@google-cloud/monitoring": "0.4.1"
+    "@google-cloud/monitoring": "1.6.1",
+    "typescript": "^3.7.3"
   },
-  "devDependencies": {
-  }
+  "devDependencies": {}
 }
